Update compile&run.cpp
Added some basic code for the execution piece. Updated executeInstruction, printCurrentState, and runMachineCode functions, and added instructionCount function.

NOTE: executeInstruction is NOT close to completed - only a few of the instructions actually have code. The rest have a "replace me" print to console because I did not program them.

Also, I'm not sure the best way to handle program inputs.... it would probably interrupt the table as is, and it wouldn't be the easiest thing to job off inputs to the printFunction since the input also sets the register value. That said, it's probably not SUPER important since I don't think the Fibonacci code requires inputs, but I can't remember